+++ Engine/Poo.h
@@ -6,15 +6,15 @@
 class Poo
 {
 public:
-	void Init( int in_x,int in_y,int in_vx,int in_vy );
+	void Init( int in_x,int in_y,double in_vx,double in_vy );
 	void Update();
 	bool TestCollision( const Dude& dude ) const;
 	void Draw( Graphics& gfx ) const;
 private:
 	int x;
 	int y;
-	int vx;
-	int vy;
+	double vx;
+	double vy;
 	static constexpr int width = 24;
 	static constexpr int height = 24;
 	bool initialized = false;
